+++ web/src/main/java/com/navercorp/pinpoint/web/applicationmap/dao/hbase/MapScanFactory.java
@@ -35,8 +35,10 @@ public class MapScanFactory {
         }
 
         // start key is replaced by end key because timestamp has been reversed
-        byte[] startKey = ApplicationMapStatisticsUtils.makeRowKey(application.getName(), application.getServiceTypeCode(), range.getTo());
-        byte[] endKey = ApplicationMapStatisticsUtils.makeRowKey(application.getName(), application.getServiceTypeCode(), range.getFrom());
+        byte[] startKey = ApplicationMapStatisticsUtils
+                .makeRowKey(application.getName(), (short) application.getServiceTypeCode(), range.getTo());
+        byte[] endKey = ApplicationMapStatisticsUtils
+                .makeRowKey(application.getName(), (short) application.getServiceTypeCode(), range.getFrom());
 
         final Scan scan = new Scan();
         scan.setCaching(this.scanCacheSize);
