+++ web/src/main/java/com/navercorp/pinpoint/web/dao/hbase/HbaseHostApplicationMapDao.java
@@ -128,7 +128,7 @@ public class HbaseHostApplicationMapDao implements HostApplicationMapDao {
     private byte[] createKey(Application parentApplication, long time) {
         Buffer buffer = new AutomaticBuffer();
         buffer.putPadString(parentApplication.getName(), HbaseTableConstants.APPLICATION_NAME_MAX_LEN);
-        buffer.putShort(parentApplication.getServiceTypeCode());
+        buffer.putShort((short) parentApplication.getServiceTypeCode());
         buffer.putLong(time);
         return buffer.getBuffer();
     }
